app/requirements.txt
flask==2.3.3
requests==2.31.0
schedule==1.2.0
pytz==2024.1
Pillow==10.2.0
astral==3.2